% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{add_dim_reduct}
\alias{add_dim_reduct}
\title{Store dimension-reduction results in DESeq2 object}
\usage{
add_dim_reduct(dds, n = Inf, family = "norm", batch = ~1)
}
\arguments{
\item{dds}{The original DESeq2 object containing all samples}

\item{n}{}

\item{family}{}

\item{batch}{}

\item{spec}{The analysis specificiation}
}
\value{

}
\description{
Calculate dimension reduction
}
\details{
Add a vst transformed assay, and a projection of the samples ont PCA space
}
\author{
Gavin Kelly
}
