% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.r
\name{bookdown_label}
\alias{bookdown_label}
\title{Multiple GT tables per chunk}
\usage{
bookdown_label(data, label = "")
}
\arguments{
\item{data}{The GT object}

\item{label}{The text that uniquely identifies this table amongst others in the chunk}
}
\value{
Th GT object (invisibly)
}
\description{
Generate multiple captions per chunk
}
\details{
To be used in a \code{GT} pipeline. Before the call to \code{gt} to
suffix the chunk label to make the caption unique, and once after
to reset the chunk label to its default
}
\author{
Gavin Kelly
}
