% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{build_dds_list}
\alias{build_dds_list}
\title{Generate subsets of DESeq2 object}
\usage{
build_dds_list(dds, spec)
}
\arguments{
\item{dds}{The original DESeq2 object containing all samples}

\item{spec}{The analysis specificiation}
}
\value{
A list of DESeq2 objects
}
\description{
Expand an analysis specification into its corresponding subset list
}
\details{
Generate a list of DESeq2 objects corresponding to the different
subsets specified
}
\author{
Gavin Kelly
}
