% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.r
\name{captioner}
\alias{captioner}
\title{Caption hyperlinking}
\usage{
captioner()
}
\value{

}
\description{
Make captioning function that generates hyperlinks
}
\details{
captioner returns a function that can be used in \code{fig.cap}.  Call
the resulting function with a string containing the caption text,
after any plot, to store a link to the pdf version of the
plot. Call the function without any arguments (e.g. in the
\code{fig.cap} argument of a chunk) to invoke the captioining mechanism
in the markdown.
}
\author{
Gavin Kelly
}
