% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{check_model}
\alias{check_model}
\title{Check model}
\usage{
check_model(dds)
}
\arguments{
\item{dds}{The original DESeq2 object containing all samples}

\item{mdl}{}

\item{coldat}{}
}
\value{

}
\description{
Check model
}
\details{
Run formula through an lm to check it
}
\author{
Gavin Kelly
}
