% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{d3_iMetric}
\alias{d3_iMetric}
\title{Interactive metric plot}
\usage{
d3_iMetric(data, width = NULL, height = NULL, elementId = NULL)
}
\description{
Generate an SVG scatterplot for rmarkdown reports where the user
can choose which dimensions to plot, and the colouring scheme.
}
