% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{d3_iScatter}
\alias{d3_iScatter}
\title{Interactive PCA plot}
\usage{
d3_iScatter(
  data,
  xy = grepl("^PC[0-9]+$", names(data)),
  aes = !xy,
  sample_names = row.names(data),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
Generate an SVG scatterplot for rmarkdown reports where the user
can choose which dimensions to plot, and the colouring scheme.
}
