% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{df2colorspace}
\alias{df2colorspace}
\title{Heatmap colour-scheme generator}
\usage{
df2colorspace(df, palette)
}
\value{

}
\description{
Heatmap colour-scheme generator
}
\details{
For each column in a dataframe, generate a sensible colour palette
for each column
}
\author{
Gavin Kelly
}
