% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{differential_MA}
\alias{differential_MA}
\title{Generate MA Plots}
\usage{
differential_MA(ddsList, caption)
}
\arguments{
\item{ddsList}{DESdemonA-generated list of DESeqDataSets}

\item{caption}{The function to use to create captions}
}
\description{
Generate MA Plots
}
\details{
MA plots of each differential genelist
}
\author{
Gavin Kelly
}
