% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{emcontrasts}
\alias{emcontrasts}
\title{Expand multiple comparisons into their contrasts}
\usage{
emcontrasts(dds, spec, extra = NULL)
}
\arguments{
\item{dds}{The original DESeq2 object containing all samples}

\item{spec}{}
}
\value{

}
\description{
Expand post-hoc comparisons
}
\details{
Use emmeans to expand keywords
}
\author{
Gavin Kelly
}
