% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.r
\name{enrichment}
\alias{enrichment}
\title{Geneset enrichment analysis}
\usage{
enrichment(dds, fun = "gseGO")
}
\arguments{
\item{fun}{}

\item{ddsList}{}

\item{showCategory}{}

\item{max_width}{}
}
\value{
gseaResult object
}
\description{
Geneset enrichment analysis
}
\details{
Do reactomePA-like analyses
}
\author{
Gavin Kelly
}
