% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{extract_assay}
\alias{extract_assay}
\title{Return an assay from a result-list}
\usage{
extract_assay(object, dataset, assay)
}
\arguments{
\item{object}{The list of results returned by get_result}

\item{dataset}{The name of the dataset you want to extract the assay from  - optional, and if omitted all datasets will be reported.}

\item{assay}{The name of the assay you want to extract from each dataset}
}
\value{

}
\description{
Return an assay from a result-list
}
\details{
Traverse an object returned by get_result, and extract a specific assay
}
\author{
Gavin Kelly
}
