% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{extract_colData}
\alias{extract_colData}
\title{Return PCA-projected data from a result-list}
\usage{
extract_colData(object, dataset)
}
\arguments{
\item{object}{The list of results returned by get_result}

\item{dataset}{The name of the dataset you want to extract the PC coordinates  - optional, and if omitted all datasets will be reported.}
}
\value{
A DataFrame of PC coordinates for each dataset specified
}
\description{
Return colData from a result-list
}
\details{
Traverse an object returned by get_result, and extract the colData. As the
comparisons and models within a dataset are all based off the same colData,
the return value is aggregated to the level of dataset.
}
\author{
Gavin Kelly
}
