% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{extract_pca_x}
\alias{extract_pca_x}
\title{Return PCA-projected data from a result-list}
\usage{
extract_pca_x(object, dataset)
}
\arguments{
\item{object}{The list of results returned by get_result}

\item{dataset}{The name of the dataset you want to extract the PC coordinates  - optional, and if omitted all datasets will be reported.}
}
\value{
A DataFrame of PC coordinates for each dataset specified
}
\description{
Return PCA-projected data from a result-list
}
\details{
Traverse an object returned by get_result, and extract the expression projected onto PC space.
The models, and comparisons within those, should all have the same projections, so the results
are aggregated to the level of 'dataset'
}
\author{
Gavin Kelly
}
