% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{extract_results}
\alias{extract_results}
\title{Return all results contained in a result list}
\usage{
extract_results(object, dataset, model, comparison)
}
\arguments{
\item{object}{The list of results returned by get_result}

\item{dataset}{The name of the dataset you want to restrict the traversal to - optional, and if omitted all datasets will be traversed.}

\item{model}{The name of the model you want to restrict the traversal to - optional, and if omitted all models will be traversed.}

\item{comparison}{The name of the comparison you want to restrict the traversal to - optional, and if omitted all comparisons will be traversed.}
}
\description{
Return all results contained in a result list
}
\details{
Traverse an object returned by get_result, and extract the DESeq2 results stored at each node.
}
\author{
Gavin Kelly
}
