% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{fit_models}
\alias{fit_models}
\title{Fit the DESeq2 models}
\usage{
fit_models(dds, ...)
}
\arguments{
\item{dds}{The original DESeq2 object containing all samples}
}
\description{
Fit the models of expression
}
\details{
Iterate through each model (stored in the 'models' metadata of a
DESeqDataSet) and expand the contrasts so each contrast gets a
separate nested level.
}
\author{
Gavin Kelly
}
