% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{fitLRT}
\alias{fitLRT}
\title{Fit LRT}
\usage{
fitLRT(dds, mdl, reduced, ...)
}
\arguments{
\item{dds}{The original DESeq2 object containing all samples}

\item{reduced}{}

\item{...}{}
}
\description{
Fit an LRT model
}
\details{
Insert the 'comparison' formula into the reduced slot
}
\author{
Gavin Kelly
}
