% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{get_result}
\alias{get_result}
\title{Generate the results for a model and comparison}
\usage{
get_result(
  dds,
  mcols = c("symbol", "entrez"),
  filterFun = IHW::ihw,
  lfcThreshold = 0,
  alpha = 0.1,
  ...
)
}
\arguments{
\item{dds}{The original DESeq2 object containing all samples}

\item{mcols}{}

\item{filterFun}{}

\item{lfcThreshold}{}

\item{...}{}
}
\value{

}
\description{
Generate results object
}
\details{
Insert results columns into mcols
}
\author{
Gavin Kelly
}
