% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{load_specs}
\alias{load_specs}
\title{Adjust a confounded batch effect}
\usage{
load_specs(file = "", context)
}
\arguments{
\item{inner}{A factor representing the variable to be recoded, e.g. the cell-line}

\item{...}{The parent factors that 'inner' is inside, e.g. the genotype of the cell-line}
}
\value{
A recoded factor that can now replace the inner (cell-line) variable in your model
}
\description{
Adjust a nested batch effect
}
\details{
When one covariate entirely predicts another, then including both in a
linear model will result in an unspecified model.  This can happen if we
have e.g. groups of cell-lines that represent a genotype.  If each of those
cell-lines receives a treatment, then we need both the group-level information to
test the interesting treatment x genotype interaction, but also the cell-line information
to remove any batch effect they represent.  To achieve this, we can recode the batch effect
so that each genotype has a common set of levels, and then we can include that batch effect
in interaction with the grouping factor.

If one group has a larger number of sub-treatments, then it will be necessary to remove
the columns of the design matrix that are all zero separately
}
\author{
Gavin Kelly
}
