% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{map_des}
\alias{map_des}
\title{Wrap datasets from DESdemonA}
\usage{
map_des(data, f, depth = "comparison", ...)
}
\arguments{
\item{.dmc}{A three-level list in the standard DESdemonA hierarchy dataset > model > comparison}

\item{.f}{The function that will be called on the output of the per_model or per_comparison.}

\item{before}{A function that will be invoked before .f}

\item{after}{A function that will be invoked after .f}
}
\value{
The return value of .f, as applied to the list of child outputs
}
\description{
Map over each component of a  DESdemonA object
}
\details{
If you want to loop through datasets that have been fitted by
DESdemonA, then use this function as a wrapped so that reports
can generate formatted output, and transformations can be applied.

This should be the first process in a pipeline, and be followed by
either a per_model or per_comparison call: the primary function .f
will be applied to the output.  For each dataset in the dds
object, the 'before' function will be called before passing things
on (and so can be used to print a header, for example), and the
'after' function will be called after (to produce captions).  All
functions have access to a variable '.dataset' that will take the
name of the current dataset
}
\author{
Gavin Kelly
}
