% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{mult_comp}
\alias{mult_comp}
\title{Mark a formula as a multiple comparison}
\usage{
mult_comp(spec, ...)
}
\arguments{
\item{spec}{}

\item{...}{}

\item{dds}{The original DESeq2 object containing all samples}
}
\value{

}
\description{
Post-hoc generator
}
\details{
Wrap a formula so that emmeans can auto-expand it
}
\author{
Gavin Kelly
}
