% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.r
\name{over_representation}
\alias{over_representation}
\title{Geneset over-representation analysis}
\usage{
over_representation(ddsList, fun, showCategory, max_width = 30)
}
\arguments{
\item{ddsList}{}

\item{fun}{}

\item{showCategory}{}

\item{max_width}{}
}
\value{
A list containing the plot object and table
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Gavin Kelly
}
