% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{ParamList}
\alias{ParamList}
\title{R6 class representing the set of parameters used in an analysis}
\description{
R6 class representing the set of parameters used in an analysis

R6 class representing the set of parameters used in an analysis
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ParamList-new}{\code{ParamList$new()}}
\item \href{#method-ParamList-set}{\code{ParamList$set()}}
\item \href{#method-ParamList-get}{\code{ParamList$get()}}
\item \href{#method-ParamList-publish}{\code{ParamList$publish()}}
\item \href{#method-ParamList-describe}{\code{ParamList$describe()}}
\item \href{#method-ParamList-clone}{\code{ParamList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamList-new"></a>}}
\if{latex}{\out{\hypertarget{method-ParamList-new}{}}}
\subsection{Method \code{new()}}{
Create a new set of parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamList$new(defaults = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{defaults}}{Named list of default values.  Names are the parameters, and the values will be their default.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object that will store all future values of analysis parameters.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamList-set"></a>}}
\if{latex}{\out{\hypertarget{method-ParamList-set}{}}}
\subsection{Method \code{set()}}{
Set the value of a parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamList$set(id, value, description = "", div = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The name of the parameter to be set.}

\item{\code{value}}{The value the parameter should taken henceforth; if missing, it will take the default value.}

\item{\code{description}}{A string describing what the purpose of the parameter is.}

\item{\code{div}}{Logical, whether to mention in the markdown report what the value has been set to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamList-get"></a>}}
\if{latex}{\out{\hypertarget{method-ParamList-get}{}}}
\subsection{Method \code{get()}}{
Get the value that the parameter is currently set to.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamList$get(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Name of the value you want to access.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamList-publish"></a>}}
\if{latex}{\out{\hypertarget{method-ParamList-publish}{}}}
\subsection{Method \code{publish()}}{
Turn the mutable object into a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamList$publish()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamList-describe"></a>}}
\if{latex}{\out{\hypertarget{method-ParamList-describe}{}}}
\subsection{Method \code{describe()}}{
Get a text description of what the setting is, and what value it currently takes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamList$describe(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Name of the value you want to access.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ParamList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ParamList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParamList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
