% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{qc_heatmap}
\alias{qc_heatmap}
\title{Data Visualisation}
\usage{
qc_heatmap(
  dds,
  pc_x = 1,
  pc_y = 2,
  family = "norm",
  title = "QC Visualisation",
  header = "\\n\\n##",
  param,
  caption = print
)
}
\arguments{
\item{dds}{The \link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet} object}

\item{pc_x}{The default choice for which principal component to
plot on the x axis}

\item{pc_y}{The default choice for which principal component to
plot on the y axis}

\item{family}{Choice of PCA method}

\item{title}{Data Visualisation}

\item{header}{The markdown prefix that should precede any section}

\item{param}{A parameter-set}

\item{caption}{The function to use to create captions}
}
\value{
A list (invisible) of ggplot2 objects
}
\description{
Generate visualisations of raw data
}
\details{
Plot heatmaps of the raw data, along with plots of the samples
projected onto various slices of the PCA space.
}
\author{
Gavin Kelly
}
