% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rbind_summary}
\alias{rbind_summary}
\title{rbind nested lists of data-frames}
\usage{
rbind_summary(x, levels = c("Dataset", "Design", "Comparison"))
}
\arguments{
\item{x}{the list of summaries}
}
\value{
The big table
}
\description{
rbind nested lists of data-frames
}
\details{
Walk through a DESdemonA hierarchy of summaries, and join them
into one big summary table
}
\author{
Gavin Kelly
}
