% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{rebase}
\alias{rebase}
\title{Rebase a factor's level}
\usage{
rebase(x, lev)
}
\arguments{
\item{x}{A factor to be rebased}

\item{lev}{The level of the factor that is to be regarded as the 'control' to which all others will be compared}
}
\value{
A factor with a new contrast attribute
}
\description{
Change a factor's reference level
}
\details{
Rather than change the order of the levels, this changes the way
the factor is parametrised, so that the levels are in the natural order
but the coefficients can reflect experimental design considerations
}
\author{
Gavin Kelly
}
