% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{recode_within}
\alias{recode_within}
\title{Recode nested factors}
\usage{
recode_within(inner, ...)
}
\arguments{
\item{inner}{A factor representing the variable to be recoded, e.g. the cell-line}

\item{...}{The parent factors that 'inner' is inside, e.g. the genotype of the cell-line}
}
\value{
A factor with recoded levels
}
\description{
Recode nested factors to avoid matrix-rank problems
}
\details{
Following the suggestion in the DESeq2 vignette, recode nested factors so that
they take common values in different clusters to avoid rank problems
}
\author{
Gavin Kelly
}
