% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{run_report}
\alias{run_report}
\title{Run DESdemonA Report on existing counts object}
\usage{
run_report(
  dds,
  spec_file,
  results = "results",
  output_file = "analyse.html",
  title = "RNASeq Analysis",
  author = Sys.info["user"]
)
}
\arguments{
\item{dds}{The DESeqDataSet object that you want to run the report
on. It needs the basic set of \code{colData} columns that are used in
the analysis plan. \code{colnames(dds)} will be used as labels in
plot, etc. In addition, if its \code{mcols} has columns of the that
are set to \code{entrez} and/or \code{symbol}, these will get added to the
report.  \code{metadata(dds)$org <- "org.Mm.eg.db"} will ensure that
the correct annotation libraries are used, but this is only strictly
necessary for functional annotation (which also requires those additional
columns in the \code{mcols} property.}

\item{spec_file}{The Analaysis Plan}

\item{results}{Directory in which to store excel results}

\item{output_file}{The name of the html report.}

\item{title}{HTML Title of the document}

\item{autor}{The name of the author to appear on the report}
}
\description{
This will generate a standard report on the DESeq2 data object you
provide it. It will store data objects in the \code{data} directory, so
that will need to be created, as will the results folder.
}
\author{
Gavin Kelly
}
