% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{subsample}
\alias{subsample}
\title{subsample}
\usage{
subsample(dds, subs)
}
\arguments{
\item{dds}{Not Sure}

\item{subs}{Not Sure}
}
\value{
Not Sure
}
\description{
Subsample
}
\details{
Think this should probably be deprecated
}
\author{
Gavin Kelly
}
