% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.r
\name{summarise_results}
\alias{summarise_results}
\title{Tabulate the size of the differential lists}
\usage{
summarise_results(dds)
}
\arguments{
\item{dds}{}
}
\description{
Tabulate genelists
}
\details{
Get genelist sizes - up, down and classed
}
\author{
Gavin Kelly
}
