% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.r
\name{tab_link_caption}
\alias{tab_link_caption}
\title{Link GT tables to a csv file}
\usage{
tab_link_caption(data, name)
}
\arguments{
\item{data}{The GT object with a caption set}

\item{name}{The filename of the csv}
}
\value{
Th GT object (invisibly)
}
\description{
Link GT tables to a csv file
}
\details{
To be used in a \code{GT} pipeline, it will store the underlying table
data in a csv file under the given name, and insert a link in the
table's caption that points to the csv file.
}
\author{
Gavin Kelly
}
