% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{table1}
\alias{table1}
\title{Table One}
\usage{
table1(dds, ddsList)
}
\arguments{
\item{dds}{The reference dds object}

\item{ddsList}{The DESdemonA list of datasets derived from the reference object}
}
\value{
a GT object
}
\description{
Produce Table 1 of sample metadata
}
\details{
Make a GT object representing all the original metadata and, for
each sample subset, the sample's presence and any extra columns introduced by a
transform.
}
\author{
Gavin Kelly
}
