% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{write_all_results}
\alias{write_all_results}
\title{Store results as text files}
\usage{
write_all_results(ddsList, dir = ".")
}
\arguments{
\item{ddsList}{A list of \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet-class()}} objects generated by DESdemonA containing results in the \code{\link[S4Vectors:Vector-class]{S4Vectors::mcols()}} slot}

\item{dir}{Directory to store the results in}
}
\description{
Store results as text files
}
\details{
Save unfiltered versions of the results in text files
}
\author{
Gavin Kelly
}
