% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{write_assay}
\alias{write_assay}
\title{Text file of assay data}
\usage{
write_assay(
  ddsList,
  assay = "vst",
  path = "results",
  formula = NULL,
  terms_to_remove = NULL
)
}
\arguments{
\item{ddsList}{A list of \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet-class()}} objects}

\item{assay}{The assay to be output (or 'norm' for normalised counts)}

\item{path}{Where to put the text files}

\item{formula}{The model that will be used to estimate the normalisation offsets}

\item{terms_to_remove}{The 'Terms' estimated from the above model that will be removed (subtracted) from values of the assay.}
}
\value{
A list of file paths to the excel files
}
\description{
Text file of assay data
}
\details{
Write all sample data, with the colData as a header
}
\author{
Gavin Kelly
}
