% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{write_results}
\alias{write_results}
\title{XLSX report of results}
\usage{
write_results(ddsList, param, dir = ".", assays = NULL)
}
\arguments{
\item{ddsList}{A list of \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet-class()}} objects generated by DESdemonA containing results in the mcols slot}

\item{param}{The parameter object used to generate the results}

\item{dir}{Directory to store the results in}
}
\value{
A list of file paths to the excel files
}
\description{
Xlsx reporting of results
}
\details{
Store all the differential gene-lists and supporting materials
in a multi-worksheet spreadsheet
}
\author{
Gavin Kelly
}
