% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{get_started}
\alias{get_started}
\title{Initiate a DESdemonA project}
\usage{
get_started(
  files = dir(system.file("templates", package = "DESdemonA")),
  dest = "."
)
}
\arguments{
\item{files}{Which files to retreive from the DESdemonA project}

\item{dest}{Where to copy the files to}
}
\value{
Nothing - purely for its side effect.
}
\description{
Populate a folder with DESdemonA starter scripts
}
\details{
To start a DESdemonA-based project, we provide a sample set
of scripts to point you in the right direction. Firstly there is
an '00_init.qmd' file that will create a universal DESeqDataSet object
from your quantified counts file.  You may need to edit this to link
it to where your quantification pipeline stores its results, and to
ensure that the full set of metadata is inserted into the colData.

The main analysis is run via "01_analyse.qmd" - you should render this
via quarto.  It will look for every '.spec' file in the current
directory.

The recommended usage is, at the start of project development,
to simply call 'DESdemonA::get_started()' in the relevant directory,
as the defaults path and files are sufficient.
}
\author{
Gavin Kelly
}
